Fix: gpencil stroke to perimeter
The stroke to perimeter function was not using the parent
matrix to account for parent transforations. This new implementation
uses a new function BKE_gpencil_stroke_to_view_space
which does both the transforamtion to parent and view space.
Additionally the tPerimeterPointList was removed and replaced
with the ListBase implementation.