Fix T84453: Crash bezier curves transfrom
When doing any proporional editing transformation operation on a bezier
curve that has hidden handles, Blender will crash (after save).

This bug is caused by the fact that during the initialization phase of the
`TransData` of the curve, hidden points are skipped in the counting.
But when the `TransData` is filled in, the hidden points are included
if proportional editing is turned on. Consequently the `td` and `tail`
pointers are missaligned (incremented) and ultimately cause a
heap-buffer-overflow.
The fix includes the hidden points in the counting to allow the flag
`TD_NOTCONNECTED` to be set for the hidden points. A potentially
better solution might be to find another way of handling the hidden
points while transforming.

Maniphest Tasks: T84453

Differential Revision: https://developer.blender.org/D10017