Fix T86968: Last UV factor in cyclic strokes
The UV factor of the last point of a cyclic stroke was using the factor of
the first point leading to unwanted sacling artifacts.

The fix sets the uv factor of the last point to the currect value (last UV
factor + length between last and first point).

Maniphest Tasks: T86968

Differential Revision: https://developer.blender.org/D10850