Fix T86947: Drag & Drop tooltip in Scene mode
The tooltip while draging a collection in Scene mode in the Outliner
was always "Link inside Collection" even if the action performed was
different. This was because the `collection_drop_init` set the
`from_collection` always to `NULL` if the Outliner display mode was
currently set to Scene.

The fix removes the check of the display mode and only sets the
`from_collection` to `NULL` if the ctrl (linking) key is held.

Maniphest Tasks: T86947

Differential Revision: https://developer.blender.org/D10864