Fix T87174: Inconsistensy skin root with extrude
The `bmo_extrude_vert_indiv_exec` was removing the `MVERT_SKIN_ROOT`
flag on the original vertices resulting in moving the skin root to the last
extruded vertex. Note that this is inconsistent with
`bmo_extrude_edge_only_exec` and `bmo_extrude_face_region_exec` which
will remove the flag on the extruded vertices.

The fix changes `bmo_extrude_vert_indiv_exec` so that the flag is
removed on the dpulicated vetex.

Maniphest Tasks: T87174

Differential Revision: https://developer.blender.org/D10908