Fix T87401: Drop down can apply the wrong modifier
The "Apply" button in the drop-down would always apply the
modifier that was currently under the mouse cursor. If the modifiers
were collapsed, that meant that it would usually apply the modifier
below the coorect one (or even no modifier at all if the cursor happend
to be right inbetween the two modifier panels).

The fix makes sure that the right modifier name is passed to the
`OBJECT_OT_modifier_apply` operator, which prevents it from using
the mouse cursor to get the modifier.

Maniphest Tasks: T87401

Differential Revision: https://developer.blender.org/D10977