Fix T87984: GPencil pressure changes not handled
The grease pencil drawing code did not handle changes in tablet pressure
if the pen position stayed the same. This causes 'jumps' in pressure
when the pen is moved after the pressure changed.

The fix solves this by checking increases in pressure while the mouse
did not move and then changing the pressure of the last point in the
buffer. This will ensure that the point currently drawn always gets
the latest pressure from the tablet.

Maniphest Tasks: T87984

Differential Revision: https://developer.blender.org/D11153