Fix T88202: Layer masks render in viewlayer
When rendering a grease pencil viewlayer that uses layer masks,
the masks would be rendered together with the layer. Since the
viewlayer option in grease pencil layers should only render the
layer specified (and no other layer) this would break the consistency.

The patch overrides the layer mask opacity for when a viewlayer with
a grease pencil layer with a mask is rendered. The opacity is then
restored after the drawing caches are created.

Maniphest Tasks: T88202

Differential Revision: https://developer.blender.org/D11234